fix: object type to int32

+++ datasets/ljspeech.py
@@ -104,7 +104,7 @@ class LJSpeech:
 
             text = self.normalize(text) + "E"
             text = [self.c2i[char] for char in text]
-            self.text_data.append(np.array(text, dtype=np.uint8).tostring())
+            self.text_data.append(np.array(text, dtype=np.int32).tostring())
 
     def save(self):
         if not os.path.exists(self.processed_path):
